Use non-dynamic JS for links#index
The JavaScript for links#index is not only unchanging for long periods
of time (ie. could be page-cached), it's actually totally static.

So instead of hitting the dynamic JavaScript controller and running
through the Rails stack, just reference a static JavaScript file on
disk. Will most likely go even further with this in the future, and just
include the content in the document "head" anyway, seeing as it is so
intimately bound to what's in the view.

Signed-off-by: Wincent Colaiuta <win@wincent.com>