Move links#index JS into document head
There is no benefit to having this JS in a separate file (separate
request, more latency) and the "unobtrusiveness" of the code is not in
any way enhanced either (no difference between referencing a script from
the head and just including it in the head).

This is the case because this code is intimately bound to its view -- it
is not something generic to be reused across other views -- so there is
no point in externalizing it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>