Move wiki article JavaScript into document head
This commit continues with the general move away from the dynamic
JavaScript controller, except where it is needed.

There are various levels of "dynamism" for JavaScript:

1. Totally dynamic (different for each request, depending on the user):
   example of this is the issues#show JS, which is different for admin
   users.
2. Mostly static but changes occasionally during running (for example,
   whenever a new product is added to the database; this would be the
   case for the issues#show JS because it has a product pop-up menu,
   but it is already "type 1"). This kind of JS could in theory be
   page-cached provided that care was taken to expire it correctly when
   needed.
3. Even more static, may change for new deployments (for example, JS
   which uses path helpers such as "tweets_path" to generate URLs;
   routing changes will normally only occur when deploying new versions
   of the software). These could be page-cached without having to worry
   about expiry because every new deployment would effectively expire
   them anyway.
4. Totally an utterly static, no matter how many times the ERB is
   re-evaluated. This kind of JS could be page-cached with total
   impunity without ever having to worry about expiration issues.
5. Totally static _and_ specifically bound to a specific view template.
   Given that this kind of JS need not be shared across views there is
   little call for having it in a separate file and it can just as well
   be included directly in the "head" of the document.

The articles#edit and articles#new JS falls into this last category.
Strictly speaking the same code _is_ used by two different templates,
but by embedding it in the "head" section we are not adding new
duplication; we are just moving it to a move sensible location (ie. next
to the view where it belongs).

This is a micro-optimization because it is saving one hit to the Rails
stack, admittedly for a page which only admins can ever access.

Is this a step backwards, to a more "obtrusive" style of JavaScript? No,
I don't think so. For me, "unobtrusive" JavaScript means not littering
your views with JS code (like tangled up spaghetti). Literally including
tightly-bound JS in the "head" instead of including it in the "head"
isn't really a problem; if anything, it becomes easier to maintain.

I did consider splitting the common code from the "edit" and "new"
actions into a separate file but there doesn't seem to be a worthwhile
gain. (You're eliminating minor repetition, two things, but if you
eliminate that repetition and want to update the code, or one of the
templates but not the other, you have just as much maintenance work
anyway, either updating both templates, or having to split the code off
for each template.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>