Teach get_parent method to handle "new" action
This allows us to accept requests to "twitter/123/comments/new" and the
like.

One thing to note about the following line:

  raise 'unexpected action' if action && action != 'new'

In practice the exception will probably never be raised because in
theory the routes won't allow other actions to come through (that is
"twitter/123/comments/edit" would never be let through, because the
correct form is "twitter/123/edit"; likewise "twitter/123/comments/foo"
would never be let through).

Signed-off-by: Wincent Colaiuta <win@wincent.com>