Set up comments#new
Up until now there has been no need for an independent comments#new
action; comment forms were always rendered within the context of their
parent resources.

With the advent of tweet comments, however, we're going to have a
problem. The tweet#show HTML is page cached, and that means we can't
embed the comment form (it is different for different users).

We will shortly implement an AJAX callback to pull down the comment form
on demand, but in order to fall back gracefully in the absence of
JavaScript we need a non-AJAX workflow as well.

This commit implements that workflow. Now visiting paths like:

  /issues/123/comments/new
  /blog/foo/comments/new
  /wiki/bar/comments/new
  /twitter/123/comments/new

All work as expected, providing an appropriate comment form which
enables the creation of new comments, as well as including a link back
to the parent resource.

Signed-off-by: Wincent Colaiuta <win@wincent.com>