Improve fidelity of the last_activity helper method
The initial version of the last_activity method just compared the
"updated_at" field of the commentable model with the "last_commented_at"
field.

This was a reasonable approximation for most cases but failed to reflect
a change in the feed when a comment was edited (because the edit would
not cause any change in the "last_commented_at" field).

We instead compared the "updated_at" of all comments rather than just
relying on the "last_commented_at" field. Note that this works nicely
because the passed in comments should only ever be "published" ones
(that is, neither private nor awaiting moderation, which by definition
should never appear in feeds).

Signed-off-by: Wincent Colaiuta <win@wincent.com>