Redirect to comments#new on failure
Now that comments#new works as an independent view we can take advantage
of this and render it in the event of an error while creating a new
comment.

Not that long ago we just threw an exception when a user submitted a
blank comment, which was not very good.

The next step was to at least show a flash message by redirecting to the
original parent model's "show" action.

This final step allows us to not only show a flash message, but also
display detailed information about why validation failed:

  1 error prohibited this comment from being saved
  There were problems with the following fields:
      * Body can't be blank

Signed-off-by: Wincent Colaiuta <win@wincent.com>