Display gray color behind navbar-bg.png
This makes things scale a bit more nicely in Safari at extreme sizes.
Firefox doesn't need this as it scales all images on the fly (tested in
Firefox 3).

Safari, on the other hand, leaves the images at a static size, thus
revealing a somewhat ugly "gap" of light gray (tested in Safari 3; I
believe that Safari 4 has the same behaviour as Firefox).

So paint in the background with dark gray to ameliorate the problem a
little.

Signed-off-by: Wincent Colaiuta <win@wincent.com>