Partial fix for "log in"/"log out" alignment problem in Firefox
Use floats for the list items rather than inline display, thus working
around the alignment discrepancy in Firefox.

Unfortanately this introduces a new problem in both browsers, and that
is that the "log in"/"log out" links are no longer right-aligned, which
doesn't look very good and also clashes with the search box.

I am going to try and fix that issue; if not, will have to revert.

Signed-off-by: Wincent Colaiuta <win@wincent.com>