Add mail icon to "contact" links on static pages and layouts
Seeing as the icon appears to the right of the link, move the link
itself (swap it with the "Copyright and legal info" link) so that the
icon dangles on the right rather than off-center in the middle. This
looks better.

At the same time insert a couple of non-breaking spaces on either side
of the separator bar to provide a more obvious visual separation of the
two elements.

Signed-off-by: Wincent Colaiuta <win@wincent.com>