Set up navbar search box unobtrusively
At the same time correct the positioning of the element. It looks like
upon adding the "About" item the box stayed where it was, thus
overlapping the text.

Switch to relative position so that the item will move if the navbar
changes again; but use a small font size and zero padding and margins to
ensure that it doesn't cause its parent list item to move up or down
when toggled.

Signed-off-by: Wincent Colaiuta <win@wincent.com>