Special-case taggables_search_summary for no existing tags
Avoid an ugly glitch when no existent tags are specified; instead of
printing "O items tagged with" (nothing), print "0 items tagged with
specified tags".

Note that the case of tags which exist but which are unused (as are
produced when you tag something and then untag it) the message is
already ok ie. it says "0 items tagged with (the tag name)".

Signed-off-by: Wincent Colaiuta <win@wincent.com>