Use explicit parameters in post helpers
It's bad form to rely on implicit (instance variable) parameters for
these helper functions, so change them all to explicit parameters.

Note that here we actually fix a bug in the comments_link method, where
we had an explicit parameter but were inadvertently relying on the
implicitly available instance variable in one place (inside the link_to
call).

Signed-off-by: Wincent Colaiuta <win@wincent.com>