Remove unused "editable" helper method
This is currently used nowhere in the app, and has perhaps never been
used. In any case the implementation is not correct seeing as it relies
on the "@user" instance variable being set when it would make more sense
to take an explicit "model" parameter and get the user from the model.

Signed-off-by: Wincent Colaiuta <win@wincent.com>