Remove excerpt_html helper
This is a companion commit to commit 7e8bea9 ("Remove body_html
helper"). There is no real need for this helper; the intention is
probably clearer for someone reading the view template if we just do the
explicit call of the "w" method with an appropriate
":base_heading_level" option.

Signed-off-by: Wincent Colaiuta <win@wincent.com>