Use post_path helper in comments_link helper
While the old-style use of link_to works in the running app, it doesn't
work from within view specs, so switch to using the post_path helper
instead.

This will be necessary when we flip the switch to make posts (and
articles too) accept comments by default, because the comments_link code
path will then be exercised in specs.

Signed-off-by: Wincent Colaiuta <win@wincent.com>