Make posts, articles and tweets accept comments by default
When the app was initially deployed a conservative approach was taken
for all models except the issues and topics models. The latter had
comments enabled by default and a UI exposed for adding new comments.

The other models, posts and articles, in theory were commentable but
there was no UI for posting and comments were not accepted by default.
When the tweet model was later added it too was conservative: it didn't
have comment support at all and when it was added the default of not
accepting comments was applied to it too.

Since then I created and enabled a commenting UI on blog posts and have
noticed that for every single new post I make I end up turning on the
comments field. Now that the article model has a commenting UI (not
deployed yet) and the tweet model has been made commentable and has
received the corresponding UI (also not deployed) I am sure the pattern
will be the same: I'll want comments turned on for everything.

So what was originally a conservative policy of precaution is now actual
an inconvenience. Switch the default for all these models over to
"accepts comments".

Signed-off-by: Wincent Colaiuta <win@wincent.com>