Manually trigger cache sweeper after moderation
The "moderate_as_ham!" method is especially designed to update a record
without modifying the "updated_at" field. This in turn means that no
sweepers are run after moderation, which means that caches are not
expired when they should be.

The quick-and-dirty fix applied here is to manually invoke the
appropriate sweeper from inside the controller.

Possibly more elegant solutions which I will explore in the future are:

  - check to see if a sweeper exists from inside "moderate_as_ham!" and
    call it from there instead of the controller
  - look at new Rails functionality to update records without touching
    their timestamps

Signed-off-by: Wincent Colaiuta <win@wincent.com>