Use record_timestamps in moderate_as_ham! method
This is a slightly more elegant solution than the manual invocation of
the cache sweeper in three separate controllers.

The only drawback is that it will only work when running Rails in single
threaded mode, but seeing as I intend to do that for the foreseeable
future it seems to be an acceptable compromise.

Signed-off-by: Wincent Colaiuta <win@wincent.com>