Fix breakage in comments controller specs
Calling comments#update and the subsequent Comment#save call broke this
spec because it triggers the update_caches callback method which itself
is quite complex and needs to do a Comment.find call, interfering with
the existing Comment.find mock expectation.

There are several ugly ways of fixing this, but the least ugly seems to
be to just forget the mock and test a side-effect instead.

For more details see:

  https://wincent.com/blog/thinking-about-switching-to-rr

Signed-off-by: Wincent Colaiuta <win@wincent.com>