Show a flash on wiki articles that haven't been touched in a year
The nature of wiki pages is that they go out of date if left untended,
which is quite likely to happen in the case of a wiki tended by only one
person and with a large number of pages.

Adding this notice might help steer people away from stale articles,
while providing an incentive to keep the more useful articles up to
date.

Signed-off-by: Wincent Colaiuta <win@wincent.com>