Use comments form partial in comments#edit template
I am not entirely convinced that I'm handling the whole polymorphic
comment paths issue in the best way here, but at least this reduces some
duplication. This will possibly improve when I investigate ticket #1325.

  https://wincent.com/issues/1325

(For reference, compare the approach used in the issues#new and
issues#edit templates, which keep the "form_for" line in the parent
template and pass the "f" variable in explicitly to the template.)

In any case, one possible cleanup to the approach used in this commit
might be to make a "comment_form_url" helper which returns a different
URL depending on whether the record is new or not.

Signed-off-by: Wincent Colaiuta <win@wincent.com>