Add CSS for screenshot thumbnails with "expand widgets"
This is an initial piece of the "lightbox" implementation. Given markup
like the following:

  <a href="full-sized-image.png">
    <img class="lightbox" title="caption" alt="caption" src="thumnail.png" />
  </a>

We are can now provide some CSS to make the thumbnails look like
a Polaroid photo (using a thin white border and a drop shadow visible in
modern browsers). If JavaScript is not available clicking on the image
will just load the full-sized image in the browser.

Using dynamic JavaScript (forthcoming commit) we can superimpose an
"expand widget" to show the full-sized screenshot in a "lightbox".

Signed-off-by: Wincent Colaiuta <win@wincent.com>