lightbox: show preloaded images on click
There is currently no special styling on the image, nor any manner of
dismissing it, but the basic ability to call up an image with a click is
now in place.

We've gotten rid of the global array of preloaded images and instead
tack the preloaded image inside its related thumbnail DOM object.

Signed-off-by: Wincent Colaiuta <win@wincent.com>