lightbox: dynamically position lightbox relative to thumbnail
Center the lightbox horizontally relative to the center of the document,
and center it vertically relative to the center of the thumbnail. We
perform bounds checking to make sure that we never use a negative "left"
value, negative "top" value, nor place the image so low as to make the
document grow in size.

Signed-off-by: Wincent Colaiuta <win@wincent.com>