lightbox: replace old image when re-showing lightbox
Note that a simple "replaceWith" doesn't work here (although the docs
indicate that you can pass in a DOM element, my tests only work when
passing in a string of HTML), so we use a "remove" followed by an
"append" operation.

This in itself isn't enough to force an update because our "image"
variable is currently effectively shared across all lightboxes (if you
look at the variable values at the time the "onclick" and "mouseenter"
handlers are defined, you'll see that they depend on the value of the
"thumbnail" variable the first time through the codepath).

Signed-off-by: Wincent Colaiuta <win@wincent.com>