lightbox: "uniquify" handling of click events
As previously noted our "thumbnail" variable was causing identical click
handlers to be installed on different lightboxes. That is, at the time
the click function was defined, "thumbnail" had a particular value (the
first seen value) that was then used for all lightboxes.

Pass an explicit "tag" parameter through to our click handler function
so that it can figure out which image it really is supposed to be
showing.

Signed-off-by: Wincent Colaiuta <win@wincent.com>