CSS: add "column" class
There is a problem with our existing "two-columns" class. The width is
specified as 50%, so setting a non-zero "padding" causes the columns to
require more than 100% of the available width and they no longer fit
side by side (the second column appears below the other one).

This is one possible solution. Here we add a "column" class that can be
nested inside the "two-column" divs to provide padding to everything
inside.

The alternative approach is to add padding to all of the elements which
might possibly appear inside the "two-column" divs. Until we try with
some real product pages we won't know which approach is going to work
out best.

Signed-off-by: Wincent Colaiuta <win@wincent.com>