Add support images for product pages
There are three main categories of images added in this commit:

  1. "Source" files which have Fireworks metadata and can be used for
     future editing.
  2. Generic images for the "public/images" directory which will be used
     by more than one product.
  3. Product-specific images for a "public/images" subfolder which will
     likely be used by only one product (in the case of this commit,
     Synergy).

While I am not so concerned about "generic" images living in the
repository I am not entirely satisfied with the storage of
product-specific images.

This application is intimately tied to the wincent.com website, and at
the moment I am the only consumer of the code, but I have tried to keep
business-specific assumptions out of the codebase as much as possible
while still tailoring the application design exactly to my requirements.

So when the source eventually goes open source it may be that somebody
tries to actually use this code as a whole rather than just borrowing
pieces of it, and in that case all of these product-specific images are
going to be an annoyance.

I'm trying to think of a way of storing them in a separate respository
and still having easy deployments. Once I've settled upon a solution
I'll remove these images from the "wincent.com" repo.

(The likely solution at this stage is going to be to create a separate
"assets.wincent.com" repo, and possibly include it as a submodule in the
"wincent.com" repo; failing that, it could be referenced via symlink and
set-up or updated as another part of the Capistrano deployment process.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>