Make "assets" subdirectory
This is a preliminary step to splitting some product-specific stuff into
a different repository.

Now all such stuff lives in "assets/public", with a symlink pointing to
it at "public/assets". Why not just link to "assets"? Because if this
becomes a different repository then it will end up having a ".git"
folder in it and we don't want that being web-accessible; in this way
only the stuff under "public" will be accessible.

The criterion for deciding whether stuff got moved into the "assets"
directory was basically: is this resource used in any application
templates? If it is only referenced by pages created by the user and
stored in the database then it really belongs in the separate "assets"
zone.

There are four interesting files which are in the "gray area"; one is the
Wincent logo, which evidently is entirely product and website-specific,
but it is for now allowed to continue on the basis that the "identity"
of the application is that it is "the software which powers
wincent.com".

The other "gray area" file is the large spinner, which is not referenced
by any application template yet, but it used by the lightbox
implementation. Given that the lightbox implementation (the JavaScript
and the CSS) is part of the application, it makes sense to keep the
spinner in place. The "product" model, after all, is an integral part of
the application, even if the actual product pages that the user provides
are not. The same arguments go for the other files used by the dashboard
implementation (the expand and close widgets).

From here a possible next step is to literally move this stuff into a
separate repo and either reference it by symlink or incorporate it as a
submodule. I think I prefer the submodule approach as it should work
identically on both local development and remote production
environments.

Signed-off-by: Wincent Colaiuta <win@wincent.com>