lightbox: fix spinner glitches
The existing spinner approach had visual glitches because of limitations
in the GIF file format (true alpha transparency is not supported). This
makes for horrible antialiasing against dark backgrounds, and makes
alternative options such as incorporating a semi-transparent background
or a drop shadow unworkable.

So make the spinner image totally opaque with a back background, and
then let the browser handle the alpha transparency when rendering.

This allows us to switch to a black background for the spinner, which is
what we wanted all along.

Signed-off-by: Wincent Colaiuta <win@wincent.com>