Use "happier" colors for product navbar buttons
The old buttons looked a little too dark, too close to black, so use a
"happier" blue for all. Switch to green background on highlight to
restore contrast between hover and non-hover states. Experimentation
showed that a green background on the right-hand navigation looked too
garish, so use a red variant instead.

Signed-off-by: Wincent Colaiuta <win@wincent.com>