Sort products by category then position
This is a case of the "Simplest Thing That Could Possibly Work". We take
advantage of the fact that we know beforehand that there will only be
three categories, "Consumer", "Developer" and "Server", and that their
alphabetical ordering is already the desired order.

(In the event that we wanted to override this order in the future we
could always add another "category_position" column to the table.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>