Group product select menu according to categories
Currently all customer-facing menus are updated (issues#new,
issues#search) as well as one admin-only menu (issues#edit) because it
it shares the same form with issues#new.

Still to do are the AJAX menus in issues#show, which might be a little
bit more tricky, so will do them as a separate step.

Closes:

  https://wincent.com/issues/1271

Signed-off-by: Wincent Colaiuta <win@wincent.com>