Remove admin_only helper
This was used only one place and it is more complicated (internally) and
longer (in keystrokes) than the alternative of simply saying "if
admin?".

It may have made sense when using another templating language, but with
Haml the simple conditional is just as concise and readable, so it
deserves to be ripped out.

Signed-off-by: Wincent Colaiuta <win@wincent.com>