Haml 2.2.0: use new abbreviated attribute syntax where appropriate
Switch to the new abbreviated syntax where appropriate.

There are a couple of places where it is not appropriate and it's been
left using the old syntax:

- places were we want to set a nil class in order to suppress the
  emission of an attribute
- places where we have to pass a method call result, and wrapping the
  call in "#{}" would actually make the line longer than or as long as
  before
- the "html_attrs" shortcut used in layouts (because there are no
  guarantees that it will work with the short syntax)

Officially the docs say that you can use bare instance and local
variables, but there are a couple of places where I use a simple method
call (with no params) and it seems to work (ie. href=root_path and the
like).

Signed-off-by: Wincent Colaiuta <win@wincent.com>