cap: enable Git submodules
This paves the way for the separation of the "assets" subdirectory into
a separate repository, which will then be included as a separate
submodule.

For the time being, it should have no real impact on deployments,
because it just causes the following commands to be run on deployment,
and they are no-ops in the absence of any submodules:

  git submodule init
  git submodule update

I haven't wanted to use submodules for other purposes at this stage
because I didn't want to introduce another point of failure (for
example, if I included RSpec or anything else inside "vendor" via
submodules) because deployments might be broken if the remote vendor
repository was off-line. The alternative would be to clone the remote
vendor's repository locally, and include that via submodule mechanism
instead, but it is awfully complicated and just freezing the desired
code into the repository is much simpler and just as effective.

Those reliability concerns don't apply to the "assets" repo, because it
will be hosted alongside the wincent.com repo itself. If you can deploy
the latter then you can by definition deploy the former too.

Signed-off-by: Wincent Colaiuta <win@wincent.com>