Add "hide_from_front_page" field to Product model
There are some "products" which are really more "projects" than
"products", and they are certainly not something "sellable" (for
example, the "wincent.git" product).

These are products which we want in the database so that we can assign
tickets to them in the issue tracker (for example), but which we don't
want cluttering up our products#index action, where we want to put our
marketable products.

Signed-off-by: Wincent Colaiuta <win@wincent.com>