Redirect products index to old site
Until more products are Snow Leopard-ready, it doesn't make sense to do
product pages for them, and that means that we only have Synergy pages
ready at the moment. It is therefore a little senseless to use the index
page, as it will only end up pointing to empty product pages.

Migration plan then is this:

  - deploy new version
  - populate Synergy pages
  - set up other products as much as possible
  - prep other products for Snow Leopard
  - as each product becomes ready, add new product pages for it
  - eventually stop redirecting the index (will require new deployment)

Signed-off-by: Wincent Colaiuta <win@wincent.com>