Ignore Wikitext Makefiles
Rails 2.3.3 introduced a change to the gem system in which a vendored
gem without a Makefile present would be considered "unbuilt". This meant
that I could no longer remove such Makefiles without breaking the server
(would refuse to start).

How do we stop this kruft file from cluttering up "git status" listings?

One option is to store the file in the repo but that's just wrong
because it is platform-dependent (changes on server from workstation)
and can and should be generated on demand.

The solution I go with here is to just add a .gitignore file with a
special glob pattern in it. I restrict it to the Wikitext gem because I
don't want to rule out adding other C-extension-based gems later on
which may depend on the Makefile being present.

Signed-off-by: Wincent Colaiuta <win@wincent.com>