script/deploy: add "wait" subcommand
In practice an "unlock" immediately followed by anything else is
unlikely to work because there appears to be some delay before the
remote server will allow unlocked accounts to actually access.

So introduce this undocumented "wait" subcommand that just waits 5
seconds, and include it in the "prep" and "all" convenience commands.

We don't want to clutter up the documentation with references to "wait"
because if you are manually working on the command line you probably
don't need any special delay.

Signed-off-by: Wincent Colaiuta <win@wincent.com>