Fix commenting on posts/articles with periods in permalinks/titles
Due to a wart in Rails' handling of nested routes with requirements on
them, posting comments on posts with periods in the permalinks or
articles with periods in the titles was broken (would raise a routing
error -- "no route matches" -- on posting).

This is because the "requirements" parameter doesn't "trickle down" once
we get to the nested part of the route.

The workaround is to explicitly repeat the requirements when declaring
the nested resource.

Fixes:

  https://wincent.com/issues/1410

Signed-off-by: Wincent Colaiuta <win@wincent.com>