script/status: avoid false positives for "running" test
It is possible that grepping for "nginx: master process" or similar will
return true because the grep process itself shows up in the list.
Mitigate this possibility by doing a two-phase grep. First we grep for
"nginx", then we grep for something that differentiates it from the grep
invocation itself. Will probably tighten this further in a subsequent
commit by doing an inverse grep ("grep -v grep").

Signed-off-by: Wincent Colaiuta <win@wincent.com>