Mark JsController routing specs as pending
With a recent update (either to Rails 2.3.4 or RSpec 1.2.9, most likely)
the JsController routing specs started failing even though the
application behaviour was correct.

In an effort to fix the breakage I dumped the old
"route_for"/"params_from" idiom and adopted the newer "route_to" one,
but the breakage continues.

So this commit marks those specs as pending until the real cause of the
breakage is identified.

See:

  https://rspec.lighthouseapp.com/projects/5645/tickets/907

Signed-off-by: Wincent Colaiuta <win@wincent.com>