Add "markup type" pop-up to page form, and use body_html method
With this commit we can now set or toggle the markup type of a given
page, and when displayed the appropriate transformation will be applied.

The way we handle the enumeration "type" here is a little different from
how it is done in the Issue model, so we may wish to examine that and
migrate to a more consistent style.

Signed-off-by: Wincent Colaiuta <win@wincent.com>