script/deploy: check that JavaScript is minimized during "check" target
Just in case the developer forgets to run "rake js:minify:create", do an
automated check during the "check" target.

Note that we use LOCAL_REF and REMOTE_REF temporary variables here to
keep things clear and not clobber the value of the REF variable set
during command line argument processing.

Due to apparent bugs in Google's Closure compiler (the command-line
version, at least), we are still using the Yahoo compressor and the
Closure support is commented out.

Another issue still to be resolved is how to actually bail in the event
of an error seeing as our "exit 1" commands don't actually abort the
shell script in the event that minimization wasn't performed correctly
(because the exit status applies to the subshell delimited by the
parentheses, and doesn't "bubble up" and terminate the actual shell
script.

Signed-off-by: Wincent Colaiuta <win@wincent.com>