script/deploy: add "bail()" function and use it
As mentioned in commit 46b52cb, there were places in the script where an
error wasn't actually aborting the run because of our use of subshells.
Even when running with the "-e" switch, constructs like this won't work:

  try_something || (echo "failed!" && exit 1)

The problem is that the "exit 1" only exits the subshell, and even
though the exit status does "bubble up" to the parent shell, it is not
enough to abort the parent script because "-e" only affects "simple
commands", not subshells.

In other words, this will abort:

  #!/bin/sh -e
  ls non-existent-directory
  echo "never shown"

But this will not:

  #!/bin/sh -e
  (exit 1)
  echo "this is shown"

The solution is to not use subshells for grouping the two statements of
the echo-then-exit pattern; rather we must use a function like the
"bail" function added in this commit.

Note that our other uses of subshells _do_ work because they _are_
considered "simple commands". For example:

  ssh $SERVER "sh -c '(test -d $A || (echo \"failed\" && exit 1)) &&
                      (test -d $B || (echo \"failed\" && exit 1)) &&
                      (test -d $C || (echo \"failed\" && exit 1))'"

Here because the commands are chained together rather than sequential,
every single one must pass with zero exit status in order for the "sh"
command to return a zero exit status, and in turn "ssh". If any of the
three tests fail, the exit status will be 1, and as "ssh" is a "simple
command", the shell script will abort.

Signed-off-by: Wincent Colaiuta <win@wincent.com>