Switch to Capybara + Culerity in favour of Webrat
Was thinking about moving to Culerity/Celerity anyway seeing as Selenium
is so fiddly, flakey and brittle, but Webrat doesn't support it so I
would have had to rewrite all features.

Then Capybara came out which aims to be API-compatible with Webrat but
work with any number of browser "drivers", including Culerity.

Turns out that it works fairly well and only minimal changes were
required.

Signed-off-by: Wincent Colaiuta <win@wincent.com>