Vendor all gems required by "cucumber" environment
Trying to stabilize the Cucumber infrastructure a bit and with Capyraba
and Culerity etc things seem to be a bit more solid than before.

Note that Culerity 0.2.3 won't work here, so we are actually running the
current HEAD of the master branch. The version number is a bit
misleading -- it is not really 0.2.4 -- but we couldn't use a number
like 0.2.3.head-sha1 because either RubyGems or Rails' own gem handling
isn't savvy enough to sort out the dependencies (because Capybara asks
for version ">= 0.2.3" and the system doesn't seem to realize that
"0.2.3.x" satisfies that condition).

Signed-off-by: Wincent Colaiuta <win@wincent.com>