Capybara: use "have_content" for text matching
We're losing some functionality here (ability to do "fuzzy" matches on
user-visible text) because the "page.body" method doesn't return the
actual HTML representation of the DOM but instead returns the HTML as it
was originally transmitte to the browser.

This is why our old features didn't work when run under Culerity,
because the effects of the JavaScript interaction were not reflected in
the "page.body".

What we really need here is the equivalent of Celerity's "browser.text",
which does give us the HTML representation of the DOM. We could "reach
in" and use a private method (specifically
"page.driver.send(:browser).text"), but that feels messy and is unlikely
to work with other drivers in a reliable way.

So we renounce that functionality and just use the "have_content"
matcher which is supported by all drivers.

Unfortunately it doesn't seem to support regular expression matches
(probably because it is using XPath under the hood), but at least it
works reliably.

Signed-off-by: Wincent Colaiuta <win@wincent.com>