script/status: eliminate false positives for "running"
As noted in commit 18115d0, tests like the following can fail (false
positive) because of the way piped processes get hooked up:

  ps x | grep nginx

"ps x" does not run before "grep nginx", but at the same time (the shell
hooks up the input and output and executes both simultaneously). This
means that sometimes "grep nginx" will appear in the process listing for
"ps x", depending on subtle timing races (like whether the system has
updated the process info tables which "ps x" accesses).

The attempt at avoiding such false positives implemented in commit
18115d0 was incomplete:

  ps x | grep nginx | grep 'nginx: master process'

Here we try to do a "two phase" grep -- first narrowing things down to
just the lines containing "nginx" and then trying to find the line we
are really intereested in -- but this suffers from exactly the same flaw
as the previous example; namely, all three processes are running
simulatneously, so the notion of "first this, then that" simply doesn't
apply.

For example it is quite possible that "ps x" in this case could list
both "grep nginx" and "grep 'nginx: master process'", or it could list
neither, or it could list one or the other, independently of whether or
not nginx is actually running.

The only real solution is the one mentioned in the commit message for
commit 18115d0, running "grep -v grep" immediately after "ps x" (note
here that "after" doesn't mean "after in time" but "next in the series of
pipes"). That way, even if "grep nginx", "grep 'nginx: master process'"
or even "grep -v grep" itself, is included in the output of "ps x" (and
there is no way we can ever know in a deterministic fashion which will
be the case), then those lines will be stripped out from the output,
leaving only the legitimate match that we really care about.

Signed-off-by: Wincent Colaiuta <win@wincent.com>