Tighten gem version requirements in "cucumber" environment
In the flakey, ever-changing would of Ruby it is a very bad idea to
specify version requirements for dependencies in terms of ">=" because
while version "x.y.z" may work fine it is perfectly possibly that
version "x.y.z+1" won't work at all and will completely break the
application.

So replace those ">=" requirements with hard "=" ones.

Signed-off-by: Wincent Colaiuta <win@wincent.com>