Add exception mailer for internal server errors
So as not to rely on third-party plug-ins, instead of installing a
plug-in like the "exception_notification" plug-in, instead just bake in
very basic support for sending email notifications of 500 "internal
server errors".

Notable omissions (that are included in the "exception_notification"
plug-in and also in Rails own built-in "rescue" pages for local requests
in development mode) are things like environment dumps, session dumps,
cleaned-up backtraces and such.

The main purpose of the notifications at this stage is just to alert
that a problem may exist; not to necessarily capture all of the
information needed to diagnose the problem.

Signed-off-by: Wincent Colaiuta <win@wincent.com>