Remove "logout"/"login" links rather than just hiding them
This makes it much easier to test using Cucumber/Capybara because the
"has_content?" matcher will return true for elements which are hidden
(JQuery leaves them in the DOM, of course, and just adds
'style="display: none;"' to them).

So we remove them from the DOM rather than hiding them, and now we can
use the Capybara "has_content?" matcher in a straightforward fashion,
just specifying "I should see"/"I should not see" rather than having to
worry about complicated XPath specification of the element we are
looking for and the style attribute we expect to find for it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>